requirements.txt
django==5.0.2
channels==4.0.0
django-cors-headers==4.3.1
websockets==12.0
vapi-python==0.1.9
python-dotenv==1.0.0
PyAudio==0.2.13
wave==0.0.2
daphne==4.1.0 